fix: add rawDateRead and rawDatePublished in the template
